BI/requirement.txt
Django>=5
djangorestframework
djangorestframework-simplejwt
psycopg2
python-decouple